simple elden ring like loading screen
A loading screen that can be used during transitions between scenes so that the game does not freeze while a new scene is loading.

This is a small loading scene that should be easy to integrate.

Or not to integrate, it's up to you to decide)

I just had the idea to add this scene after your video, because projects do not always require such complex checks of resource loading.